# Significantly Accelerated K-means Clustering via Parallelism Enhancement