## What is this?

This is a optimization for `nuq.py`, the main clustering step of SqueezeLLM.

## Background
The [KMeans implementation of scikit-learn](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.KMeans.html) automatically utilizes all CPU cores, on the basis of the environmental variable `OMP_NUM_THREADS`.

The default value is equal to the number of cores on the machine, so in the current state, every K-means invocation attempts to utilize all cores.

## Observation

Each K-means task in the quantization process is relatively small.

You only run one task per row, with 1-dimensional values and a small number of clusters (i.e. 8 for 3-bit and 16 for 4-bit).

Therefore it is not ideal to have all cores working on each task.

A better strategy for CPU utilization is to offload each task to a separate core, thereby handling each task with just a single thread.

## Implementation

I used the `Pool` construct from Python's `multiprocessing` library to distribute rows to a pool of K-means workers.

One important necessary change is that each K-means task needs to use only a single core - to achieve this, we must set `OMP_NUM_THREADS` to 1. There are many ways to achieve this, as goes for environmental variables in general, but I decided to include in the script `os.environ["OMP_NUM_THREADS"] = "1"`, so that you don't need to think about it when you run the Python script.

If you don't like this, you may drop that line and prefix your command with `OMP_NUM_THREADS=1` whenever you run `nuq.py`, but note that failure to do so will result in drastic performance drops, as you end up having a pool of workers each trying to use all cores.

To accommodate the multiprocessing pool, a necessary preprocessing step to generate the tasks and a postprocessing step to collect the clustering results have been added.

## Results

All numbers were measured on a i9-13900K machine with 32 threads.

For the 4-bit quantization on **Llama-2-7B**, the original runtime per layer stood at 4 minutes and 10 seconds, or **250 seconds**.

My implementation, which generates identical results with no tradeoffs, achieves the same in **11 seconds**.

**This is a 22.7-fold speedup, and Llama-2-7B can now be quantized in under 6 minutes, as opposed to the 2+ hours previously required.**

Comparable results have also been observed with other models of varying size, at both 3 and 4-bit quantizations.

## Further suggestions

As the time it takes for each layer decreases dramatically with this improvement, I think it may be better to move the `tqdm` progress bar up in the quantization process so that it runs only once, with a tick for every layer. A progress bar every 11 seconds may seem like too much.

Further optimizations, including custom K-means implementations exploiting the 1D nature of the task and parallelization using [Numba](https://numba.pydata.org/) proved to also show notable gains in my experimentations, but these require extensive modifications of code, so I do not plan on PRing these to this repository. I point this out so that if runtime is an issue, you may look into these.