@SyphonArch 2024-03-22
Note that this parallelism change affects the K-means random_state reproducibility - the previous configuration and this configuration don't have compatible seeds. Other than that, they will produce identical results.