@kssteven418 2024-03-22
After merging the PR, I tested on my end that the numbers for both dense-only and dense-and-sparse are reproducible.

Thank you for the great PR and this really enhanced the useability of the framework!